requirements.txt
# requirements.txt
django>=4.0,<5.0  # Django framework
djangorestframework>=3.14,<3.16  # Django REST Framework
psycopg2-binary>=2.9,<2.10  # PostgreSQL adapter (agar PostgreSQL ishlatsangiz)
djangorestframework-simplejwt>=5.0,<5.4 # JWT autentifikatsiyasi uchun
python-dotenv>=0.20,<1.2 # .env faylidan sozlamalarni o'qish uchun
django-parler>=2.3,<2.4
django_cors_headers
python-telegram-bot[httpx]>=20.0,<21.0
requests==2.32.3
django-debug-toolbar>=3.8,<4.4